+++ src/unpoly/element.js
@@ -740,7 +740,7 @@ up.element = (function() {
 
   Due to quirks in the `DOMParser` spec, `<script>` and `<noscript>`
   elements in the returned document will be inert. To make them active,
-  use `up.element.fixScriptish()`.
+  use `up.element.fixParserDamage()`.
 
   @function up.element.createBrokenDocumentFromHTML
   @param {string} html
@@ -761,19 +761,19 @@ up.element = (function() {
      This confuses libraries that work with <noscript> tags, such as lazysizes.
   2. <script> elements are inert and will not run code when inserted into the main `document`.
 
-  @function up.element.fixScriptish
+  @function up.element.fixParserDamage
   @param {Element} scriptish
     A `<script>` or `<noscript>` element.
   @internal
   */
-  function fixScriptish(scriptish) {
-    let clone = document.createElement(scriptish.tagName)
-    for (let { name, value } of scriptish.attributes) {
-      clone.setAttribute(name, value)
-    }
-
-    clone.textContent = scriptish.innerHTML
+  function fixParserDamage(scriptish) {
+    // We cannot use `scriptish.cloneNode(true)` as this does not fix broken <noscript> elements
+    let clone = createFromHTML(scriptish.outerHTML)
     scriptish.replaceWith(clone)
+  }
+
+  function disableScript(scriptElement) {
+    scriptElement.type = 'up-disabled-script'
   }
 
   /*-
@@ -1412,7 +1412,7 @@ up.element = (function() {
     attrSelector, // internal
     tagName: elementTagName,
     createBrokenDocumentFromHTML, // internal
-    fixScriptish,
+    fixParserDamage,
     createFromHTML, // practical for element creation
     get root() { return getRoot() }, // internal
     paint, // internal
@@ -1446,5 +1446,6 @@ up.element = (function() {
     parseSelector,
     filteredQuery,
     isEmpty,
+    disableScript,
   }
 })()